Move `verifyCsharpImplicitThrow` to own Csharp test class
